+++ changelog.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.1]
+### Added
+- Message to RouteNotFoundException and MethodNotAllowedException.
+
 ## [0.5.0]
 ### Fixed
 - MethodNotAllowedException during Router match when a wildcard route exists.